revert commit: inside "administration" (Services), use a tree. eg: CRIR can now be found in "Secrétariat Général" > "Services Centraux"
this commit is quite wrong since for example service_SG58 is inside service_SG5 but nobody has supannEntiteAffectation=SG5,
breaking the asumption of PAGS group that parent group must match.

a solution would be to use a "or" on members for non-leaf groups:
eg instead of supannEntiteAffectation=SG5, have (|(supannEntiteAffectation=SG58)(supannEntiteAffectation=SG57)...)
but is that ok/enough?

git-svn-id: http://subversion.cru.fr/esup-tools/PAGScreator/trunk@62 93fd7c73-ba33-0410-a30b-86d151360d8f